Move item, catalog, and collection to core
