+++ pystac/layout.py
@@ -8,9 +8,9 @@ from pystac.utils import safe_urlparse, join_path_or_url, JoinType
 
 if TYPE_CHECKING:
     from pystac.stac_object import STACObject as STACObject_Type
-    from pystac.catalog import Catalog as Catalog_Type
-    from pystac.collection import Collection as Collection_Type
-    from pystac.item import Item as Item_Type
+    from pystac.core import Catalog as Catalog_Type
+    from pystac.core import Collection as Collection_Type
+    from pystac.core import Item as Item_Type
 
 
 class TemplateError(Exception):
