+++ pystac/stac_io.py
@@ -32,7 +32,7 @@ except ImportError:
 
 if TYPE_CHECKING:
     from pystac.stac_object import STACObject as STACObject_Type
-    from pystac.catalog import Catalog as Catalog_Type
+    from pystac.core import Catalog as Catalog_Type
     from pystac.link import Link as Link_Type
